feat: add one endpoint to generate a valid token
