--- apps/payment-service/src/data-wrapper/data-wrapper.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DataWrapperController } from './data-wrapper.controller';
-import { DataWrapperService } from './data-wrapper.service';
-
-@Module({
-  controllers: [DataWrapperController],
-  providers: [DataWrapperService],
-})
-export class DataWrapperModule {}
