+++ apps/payment-service/src/stripe/stripe.module.ts
@@ -1,10 +1,8 @@
 import { DynamicModule, Module, Provider } from '@nestjs/common';
-import { StripeController } from './stripe.controller';
 import { StripeService } from './stripe.service';
 import Stripe from 'stripe';
 
 @Module({
-  controllers: [StripeController],
   providers: [StripeService],
 })
 export class StripeModule {
