+++ apps/payment-service/src/token-generator/dto/generate-jwt.dto.ts
@@ -26,8 +26,4 @@ export class GenerateJwtDto {
 
   @IsEnum(CurrencyTypes)
   currency: CurrencyTypes;
-
-  @IsString()
-  @IsNotEmpty()
-  client_secret: string;
 }
